+++ packages/cojson-storage-indexeddb/src/index.test.ts
@@ -3,7 +3,6 @@ import {
     ControlledAgent,
     LocalNode,
     WasmCrypto,
-    cojsonInternals,
 } from "cojson";
 import { IDBStorage } from "./index.js";
 
@@ -14,7 +13,7 @@ test.skip("Should be able to initialize and load from empty DB", async () => {
 
     const node = new LocalNode(
         new ControlledAgent(agentSecret, Crypto),
-        cojsonInternals.newRandomSessionID(Crypto.getAgentID(agentSecret)),
+        Crypto.newRandomSessionID(Crypto.getAgentID(agentSecret)),
         Crypto,
     );
 
@@ -34,7 +33,7 @@ test("Should be able to sync data to database and then load that from a new node
 
     const node1 = new LocalNode(
         new ControlledAgent(agentSecret, Crypto),
-        cojsonInternals.newRandomSessionID(Crypto.getAgentID(agentSecret)),
+        Crypto.newRandomSessionID(Crypto.getAgentID(agentSecret)),
         Crypto,
     );
 
@@ -54,7 +53,7 @@ test("Should be able to sync data to database and then load that from a new node
 
     const node2 = new LocalNode(
         new ControlledAgent(agentSecret, Crypto),
-        cojsonInternals.newRandomSessionID(Crypto.getAgentID(agentSecret)),
+        Crypto.newRandomSessionID(Crypto.getAgentID(agentSecret)),
         Crypto,
     );