+++ packages/jazz-browser-auth-clerk/src/index.ts
@@ -1,11 +1,24 @@
 import { Account, AuthMethod, AuthResult, ID } from "jazz-tools";
-import type { LoadedClerk } from '@clerk/types';
 import { AgentSecret } from "cojson";
+
+export type MinimalClerkClient = {
+    user: {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        unsafeMetadata: Record<string, any>;
+        fullName: string | null;
+        username: string | null;
+        id: string;
+        update: (args: {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            unsafeMetadata: Record<string, any>;
+        }) => Promise<unknown>;
+    } | null | undefined;
+}
 
 export class BrowserClerkAuth implements AuthMethod {
     constructor(
         public driver: BrowserClerkAuth.Driver,
-        private readonly clerkClient: LoadedClerk,
+        private readonly clerkClient: MinimalClerkClient
     ) {}
 
     async start(): Promise<AuthResult> {