+++ packages/jazz-browser/src/index.ts
@@ -11,14 +11,15 @@ import {
     CryptoProvider,
     AuthMethod,
     createJazzContext,
+    AnonymousJazzAgent,
 } from "jazz-tools";
 import { RawAccountID, LSMStorage } from "cojson";
 import { OPFSFilesystem } from "./OPFSFilesystem.js";
 import { IDBStorage } from "cojson-storage-indexeddb";
 import { createWebSocketPeer } from "cojson-transport-ws";
-export { BrowserDemoAuth } from './auth/DemoAuth.js'
-export { BrowserPasskeyAuth } from './auth/PasskeyAuth.js'
-export { BrowserPassphraseAuth } from './auth/PassphraseAuth.js'
+export { BrowserDemoAuth } from "./auth/DemoAuth.js";
+export { BrowserPasskeyAuth } from "./auth/PasskeyAuth.js";
+export { BrowserPassphraseAuth } from "./auth/PassphraseAuth.js";
 
 /** @category Context Creation */
 export type BrowserContext<Acc extends Account> = {
@@ -27,65 +28,96 @@ export type BrowserContext<Acc extends Account> = {
     done: () => void;
 };
 
-/** @category Context Creation */
-export async function createJazzBrowserContext<Acc extends Account>({
-    auth,
-    AccountSchema = Account as unknown as CoValueClass<Acc> & {
-        fromNode: (typeof Account)["fromNode"];
-    },
-    peer: peerAddr,
-    reconnectionTimeout: initialReconnectionTimeout = 500,
-    storage = "indexedDB",
-    crypto: customCrypto,
-}: {
+export type BrowserGuestContext = {
+    guest: AnonymousJazzAgent;
+    done: () => void;
+};
+
+export type BrowserContextOptions<Acc extends Account> = {
     auth: AuthMethod;
     AccountSchema: CoValueClass<Acc> & {
         fromNode: (typeof Account)["fromNode"];
     };
+} & BaseBrowserContextOptions;
+
+export type BaseBrowserContextOptions = {
     peer: `wss://${string}` | `ws://${string}`;
     reconnectionTimeout?: number;
     storage?: "indexedDB" | "singleTabOPFS";
     crypto?: CryptoProvider;
-}): Promise<BrowserContext<Acc>> {
-    const crypto = customCrypto || (await WasmCrypto.create());
+};
+
+/** @category Context Creation */
+export async function createJazzBrowserContext<Acc extends Account>(
+    options: BrowserContextOptions<Acc>,
+): Promise<BrowserContext<Acc>>;
+export async function createJazzBrowserContext(
+    options: BaseBrowserContextOptions,
+): Promise<BrowserGuestContext>;
+export async function createJazzBrowserContext<Acc extends Account>(
+    options: BrowserContextOptions<Acc> | BaseBrowserContextOptions,
+): Promise<BrowserContext<Acc> | BrowserGuestContext>
+export async function createJazzBrowserContext<Acc extends Account>(
+    options: BrowserContextOptions<Acc> | BaseBrowserContextOptions,
+): Promise<BrowserContext<Acc> | BrowserGuestContext> {
+    const crypto = options.crypto || (await WasmCrypto.create());
 
     const firstWsPeer = createWebSocketPeer({
-        websocket: new WebSocket(peerAddr),
-        id: peerAddr + "@" + new Date().toISOString(),
+        websocket: new WebSocket(options.peer),
+        id: options.peer + "@" + new Date().toISOString(),
         role: "server",
     });
     let shouldTryToReconnect = true;
 
-    let currentReconnectionTimeout = initialReconnectionTimeout;
+    let currentReconnectionTimeout = options.reconnectionTimeout || 500;
 
     function onOnline() {
         console.log("Online, resetting reconnection timeout");
-        currentReconnectionTimeout = initialReconnectionTimeout;
+        currentReconnectionTimeout = options.reconnectionTimeout || 500;
     }
 
     window.addEventListener("online", onOnline);
 
-    const { account, done } = await createJazzContext({
-        AccountSchema,
-        auth,
-        crypto: await WasmCrypto.create(),
-        peersToLoadFrom: [
-            storage === "indexedDB"
-                ? await IDBStorage.asPeer()
-                : await LSMStorage.asPeer({
-                      fs: new OPFSFilesystem(crypto),
-                      // trace: true,
-                  }),
-            firstWsPeer,
-        ],
-        sessionProvider: provideBroswerLockSession,
-    });
+    const context =
+        "auth" in options
+            ? await createJazzContext({
+                  AccountSchema: options.AccountSchema,
+                  auth: options.auth,
+                  crypto: await WasmCrypto.create(),
+                  peersToLoadFrom: [
+                      options.storage === "indexedDB"
+                          ? await IDBStorage.asPeer()
+                          : await LSMStorage.asPeer({
+                                fs: new OPFSFilesystem(crypto),
+                                // trace: true,
+                            }),
+                      firstWsPeer,
+                  ],
+                  sessionProvider: provideBroswerLockSession,
+              })
+            : await createJazzContext({
+                  crypto: await WasmCrypto.create(),
+                  peersToLoadFrom: [
+                      options.storage === "indexedDB"
+                          ? await IDBStorage.asPeer()
+                          : await LSMStorage.asPeer({
+                                fs: new OPFSFilesystem(crypto),
+                                // trace: true,
+                            }),
+                      firstWsPeer,
+                  ],
+              });
+
+    const node =
+        "account" in context
+            ? context.account._raw.core.node
+            : context.agent.node;
 
     async function websocketReconnectLoop() {
         while (shouldTryToReconnect) {
             if (
-                Object.keys(account._raw.core.node.syncManager.peers).some(
-                    (peerId) => peerId.includes(peerAddr),
+                Object.keys(node.syncManager.peers).some((peerId) =>
+                    peerId.includes(options.peer),
                 )
             ) {
                 // TODO: this might drain battery, use listeners instead
@@ -114,10 +146,10 @@ export async function createJazzBrowserContext<Acc extends Account>({
                     );
                 });
 
-                account._raw.core.node.syncManager.addPeer(
+                node.syncManager.addPeer(
                     createWebSocketPeer({
-                        websocket: new WebSocket(peerAddr),
-                        id: peerAddr + "@" + new Date().toISOString(),
+                        websocket: new WebSocket(options.peer),
+                        id: options.peer + "@" + new Date().toISOString(),
                         role: "server",
                     }),
                 );
@@ -127,14 +159,23 @@ export async function createJazzBrowserContext<Acc extends Account>({
 
     void websocketReconnectLoop();
 
-    return {
-        me: account,
-        done: () => {
-            shouldTryToReconnect = false;
-            window.removeEventListener("online", onOnline);
-            done();
-        },
-    };
+    return "account" in context
+        ? {
+              me: context.account,
+              done: () => {
+                  shouldTryToReconnect = false;
+                  window.removeEventListener("online", onOnline);
+                  context.done();
+              },
+          }
+        : {
+              guest: context.agent,
+              done: () => {
+                  shouldTryToReconnect = false;
+                  window.removeEventListener("online", onOnline);
+                  context.done();
+              },
+          };
 }
 
 /** @category Auth Providers */
@@ -142,7 +183,10 @@ export type SessionProvider = (
     accountID: ID<Account> | AgentID,
 ) => Promise<SessionID>;
 
-export function provideBroswerLockSession(accountID: ID<Account> | AgentID) {
+export function provideBroswerLockSession(
+    accountID: ID<Account> | AgentID,
+    crypto: CryptoProvider,
+) {
     let sessionDone!: () => void;
     const donePromise = new Promise<void>((resolve) => {
         sessionDone = resolve;
@@ -166,7 +210,7 @@ export function provideBroswerLockSession(accountID: ID<Account> | AgentID) {
 
                         const sessionID =
                             localStorage[accountID + "_" + idx] ||
-                            cojsonInternals.newRandomSessionID(
+                            crypto.newRandomSessionID(
                                 accountID as RawAccountID | AgentID,
                             );
                         localStorage[accountID + "_" + idx] = sessionID;