+++ packages/jazz-react/src/auth/auth.ts
@@ -1,13 +1,3 @@
-import React from "react";
-import { AuthMethod } from "jazz-tools";
-
-export type AuthState = "loading" | "ready" | "signedIn";
-
-/** @category Auth Methods */
-export const AuthMethodCtx = React.createContext<AuthMethod | undefined>(
-    undefined,
-);
-
-export { DemoAuth } from "./DemoAuth.js";
-export { PasskeyAuth, usePasskeyAuth } from "./PasskeyAuth.js";
-export { PassphraseAuth } from "./PassphraseAuth.js";
+export { useDemoAuth, DemoAuthBasicUI } from "./DemoAuth.js";
+export { usePasskeyAuth, PasskeyAuthBasicUI } from "./PasskeyAuth.js";
+export { usePassphraseAuth, PassphraseAuthBasicUI } from "./PassphraseAuth.js";