+++ packages/jazz-react/src/index.tsx
@@ -1,5 +1,7 @@
-import React, { useContext, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import {
+    BrowserContext,
+    BrowserGuestContext,
     consumeInviteLinkFromWindowLocation,
     createJazzBrowserContext,
 } from "jazz-browser";
@@ -7,6 +9,8 @@ import {
 import {
     Account,
     AccountClass,
+    AnonymousJazzAgent,
+    AuthMethod,
     CoValue,
     CoValueClass,
     DeeplyLoaded,
@@ -14,41 +18,48 @@ import {
     ID,
     subscribeToCoValue,
 } from "jazz-tools";
-import { AuthMethodCtx } from "./auth/auth.js";
 
 /** @category Context & Hooks */
 export function createJazzReactApp<Acc extends Account>({
     AccountSchema = Account as unknown as AccountClass<Acc>,
 }: {
     AccountSchema?: AccountClass<Acc>;
-} = {}): JazzReactContext<Acc> {
-    const JazzContext = React.createContext<Acc | undefined>(undefined);
+} = {}): JazzReactApp<Acc> {
+    const JazzContext = React.createContext<
+        BrowserContext<Acc> | BrowserGuestContext | undefined
+    >(undefined);
 
     function Provider({
         children,
+        auth,
         peer,
         storage,
     }: {
         children: React.ReactNode;
+        auth: AuthMethod | "guest";
         peer: `wss://${string}` | `ws://${string}`;
         storage?: "indexedDB" | "singleTabOPFS";
     }) {
-        const [me, setMe] = useState<Acc | undefined>();
-        const auth = useContext(AuthMethodCtx);
-
-        if (!auth) {
-            throw new Error("Jazz.Provider must be used within an Auth Method Provider");
-        }
+        const [ctx, setCtx] = useState<
+            BrowserContext<Acc> | BrowserGuestContext | undefined
+        >();
 
         useEffect(() => {
-            const promiseWithDoneCallback = createJazzBrowserContext<Acc>({
-                AccountSchema,
-                auth,
-                peer,
-                storage,
-            }).then(({ me, done }) => {
-                setMe(me);
-                return done;
+            const promiseWithDoneCallback = createJazzBrowserContext<Acc>(
+                auth === "guest"
+                    ? {
+                          peer,
+                          storage,
+                      }
+                    : {
+                          AccountSchema,
+                          auth: auth,
+                          peer,
+                          storage,
+                      },
+            ).then((context) => {
+                setCtx(context);
+                return context.done;
             });
 
             return () => {
@@ -56,10 +67,12 @@ export function createJazzReactApp<Acc extends Account>({
             };
         }, [AccountSchema, auth, peer, storage]);
 
-        return <JazzContext.Provider value={me}>{children}</JazzContext.Provider>;
+        return (
+            <JazzContext.Provider value={ctx}>{children}</JazzContext.Provider>
+        );
     }
 
-    function useAccount(): { me: Acc | undefined };
+    function useAccount(): { me: Acc };
     function useAccount<D extends DepthsIn<Acc>>(
         depth: D,
     ): { me: DeeplyLoaded<Acc, D> | undefined };
@@ -68,15 +81,35 @@ export function createJazzReactApp<Acc extends Account>({
     ): { me: Acc | DeeplyLoaded<Acc, D> | undefined } {
         const context = React.useContext(JazzContext);
 
+        if (!context) {
+            throw new Error("useAccount must be used within a JazzProvider");
+        }
+
+        if (!("me" in context)) {
+            throw new Error(
+                "useAccount can't be used in a JazzProvider with auth === 'guest' - consider using useAccountOrGuest()",
+            );
+        }
+
         const me = useCoState<Acc, D>(
-            context?.constructor as CoValueClass<Acc>,
-            context?.id,
+            context?.me.constructor as CoValueClass<Acc>,
+            context?.me.id,
             depth,
         );
 
         return {
-            me: depth === undefined ? me || context : me,
+            me: depth === undefined ? me || context.me : me,
         };
+    }
+
+    function useAccountOrGuest(): Acc | AnonymousJazzAgent {
+        const context = React.useContext(JazzContext);
+
+        if (!context) {
+            throw new Error("useAccountOrGuest must be used within a JazzProvider");
+        }
+
+        return "me" in context ? context.me : context.guest;
     }
 
     function useCoState<V extends CoValue, D>(
@@ -88,15 +121,19 @@ export function createJazzReactApp<Acc extends Account>({
         const [state, setState] = useState<{
             value: DeeplyLoaded<V, D> | undefined;
         }>({ value: undefined });
-        const me = React.useContext(JazzContext);
+        const context = React.useContext(JazzContext);
+
+        if (!context) {
+            throw new Error("useCoState must be used within a JazzProvider");
+        }
 
         useEffect(() => {
-            if (!id || !me) return;
+            if (!id) return;
 
-            return subscribeToCoValue(Schema, id, me, depth, (value) => {
+            return subscribeToCoValue(Schema, id, "me" in context ? context.me : context.guest, depth, (value) => {
                 setState({ value });
             });
-        }, [Schema, id, me]);
+        }, [Schema, id, context]);
 
         return state.value;
     }
@@ -110,12 +147,21 @@ export function createJazzReactApp<Acc extends Account>({
         onAccept: (projectID: ID<V>) => void;
         forValueHint?: string;
     }): void {
-        const me = React.useContext(JazzContext);
+        const context = React.useContext(JazzContext);
+
+        if (!context) {
+            throw new Error("useAcceptInvite must be used within a JazzProvider");
+        }
+
+        if (!("me" in context)) {
+            throw new Error("useAcceptInvite can't be used in a JazzProvider with auth === 'guest'.");
+        }
+
         useEffect(() => {
-            if (!me) return;
+            if (!context) return;
 
             const result = consumeInviteLinkFromWindowLocation({
-                as: me,
+                as: context.me,
                 invitedObjectSchema,
                 forValueHint,
             });
@@ -131,23 +177,25 @@ export function createJazzReactApp<Acc extends Account>({
     return {
         Provider,
         useAccount,
+        useAccountOrGuest,
         useCoState,
         useAcceptInvite,
     };
 }
 
 /** @category Context & Hooks */
-export interface JazzReactContext<Acc extends Account> {
+export interface JazzReactApp<Acc extends Account> {
     /** @category Provider Component */
     Provider: React.FC<{
         children: React.ReactNode;
+        auth: AuthMethod | "guest";
         peer: `wss://${string}` | `ws://${string}`;
         storage?: "indexedDB" | "singleTabOPFS";
     }>;
 
     /** @category Hooks */
     useAccount(): {
-        me: Acc | undefined;
+        me: Acc;
     };
     /** @category Hooks */
     useAccount<D extends DepthsIn<Acc>>(
@@ -156,6 +204,8 @@ export interface JazzReactContext<Acc extends Account> {
         me: DeeplyLoaded<Acc, D> | undefined;
     };
 
+    /** @category Hooks */
+    useAccountOrGuest(): Acc | AnonymousJazzAgent;
     /** @category Hooks */
     useCoState<V extends CoValue, D>(
         // eslint-disable-next-line @typescript-eslint/no-explicit-any