+++ examples/chat-clerk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # jazz-example-chat
 
+## 0.0.80-guest-auth.3
+
+### Patch Changes
+
+- Updated dependencies
+  - jazz-react@0.7.35-guest-auth.3
+  - jazz-tools@0.7.35-guest-auth.3
+  - jazz-react-auth-clerk@0.7.33-guest-auth.3
+
 ## 0.0.80-unique.2
 
 ### Patch Changes