+++ examples/pets/CHANGELOG.md
@@ -1,5 +1,14 @@
 # jazz-example-pets
 
+## 0.0.100-guest-auth.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   jazz-react@0.7.35-guest-auth.3
+    -   jazz-tools@0.7.35-guest-auth.3
+    -   jazz-browser-media-images@0.7.35-guest-auth.3
+
 ## 0.0.100-unique.2
 
 ### Patch Changes