+++ packages/jazz-browser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # jazz-browser
 
+## 0.7.35-guest-auth.3
+
+### Patch Changes
+
+-   Implement guest auth without account
+-   Updated dependencies
+    -   jazz-tools@0.7.35-guest-auth.3
+
 ## 0.7.35-unique.2
 
 ### Patch Changes