+++ packages/jazz-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # jazz-react
 
+## 0.7.35-guest-auth.3
+
+### Patch Changes
+
+-   Implement guest auth without account
+-   Updated dependencies
+    -   jazz-browser@0.7.35-guest-auth.3
+    -   jazz-tools@0.7.35-guest-auth.3
+
 ## 0.7.35-unique.2
 
 ### Patch Changes